+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add new FamilyInstance node - FamilyInstance.ByHostAndPoint.
+
 ## 0.3.2
 * Add new selection node  - "Select Reference on Element" which can select a reference on an element (contains linked element).
 * Add new Viewport nodes - Viewport.LabelOffset, Viewport.SetLabelOffset, Viewport.LabelLineLength, Viewport.SetLabelLineLength.